Factor out a common TransformDialectPassBase
This includes the logic necessary for injecting transforms from a file,
maintaining those alive in a separate module, debug targeting via
attributes, and transform module identification in the IR. It is
currently used for IREE codegen, and will be reused for Flow and
standalone interpreter.
