+++ llvm-external-projects/iree-dialects/lib/Dialect/LinalgTransform/Passes/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_mlir_library(IREELinalgTransformDialectPasses
   ExpertExpansion.cpp
   TransformInterpreter.cpp
+  TransformInterpreterPassBase.cpp
 
   DEPENDS
   mlir-headers
